pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mo-stock-selection"
version = "0.1.0"
description = "A 股每日批量选股系统：5 维度规则快筛 + Claude AI 深度分析"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "mo", email = "xiaomingtx18@gmail.com" }]
license = { text = "MIT" }

# 运行时依赖
dependencies = [
    # 行情数据
    "tushare>=1.4.0",
    "pandas>=2.0",
    "numpy>=1.24",

    # 数据库 / ORM / 迁移
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",

    # 配置 / Schema
    "pydantic>=2.6",
    "pydantic-settings>=2.2",

    # AI
    "anthropic>=0.40",

    # 调度 / 重试 / 日志
    "apscheduler>=3.10",
    "tenacity>=8.2",
    "loguru>=0.7",

    # CLI / 模板 / 工具
    "click>=8.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
# 开发与测试依赖：pip install -e ".[dev]"
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "pytest-postgresql>=5.0",
    "freezegun>=1.4",
    "mypy>=1.9",
    "pylint>=3.1",
    "ruff>=0.3",
    "types-pyyaml",
]

[project.scripts]
# 命令行入口：mo-stock run-once / backfill / init-db ...
mo-stock = "mo_stock.cli:cli"

# -------------------- 包发现 --------------------
# 同时识别顶层的 `config/` 和 `src/mo_stock/` 两个包
[tool.setuptools]
package-dir = { "" = ".", "mo_stock" = "src/mo_stock" }

[tool.setuptools.packages.find]
where = [".", "src"]
include = ["config*", "mo_stock*"]
exclude = ["tests*", "alembic*", "data*"]

# -------------------- 静态检查 --------------------
[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = false
ignore_missing_imports = true  # tushare / anthropic 等第三方缺类型时不报错

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4"]
ignore = ["E501"]  # 行长由 black 管

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q --strict-markers"
markers = [
    "integration: 集成测试（需要 Postgres 容器）",
    "slow: 运行较慢的测试",
]
